refactor: extract UserAppAttrs type to shared file
- Move UserAppAttrs type definition to packages/ui/src/types/user-app-attrs.ts
- Remove duplicate type definitions from restrict-client.tsx and restrict-server.tsx
- Import shared type in both components to follow DRY principles
- Improves maintainability and reduces code duplication
