+++ packages/ui/src/types/user-app-attrs.ts
@@ -1,0 +1,9 @@
+export type UserAppAttrs<TUserType = string, TUserRole = string> = {
+  admin: boolean;
+  type: TUserType;
+  slug?: string;
+  ids?: Array<{
+    id: string;
+    userRoles: TUserRole[];
+  }>;
+};
